added rsvp, refactored firestore triggers, marked old code as deprecated
